+++ src/main/scala/com/github/frankivo/actors/Telegram.scala
@@ -1,10 +1,11 @@
-package com.github.frankivo
+package com.github.frankivo.actors
 
 import java.io.File
 import java.nio.file.Paths
 import java.time.LocalDate
 
 import akka.actor.Actor
+import com.github.frankivo.CovidBot
 import com.github.frankivo.messages.{RequestCasesForDate, TelegramMessage, UpdateAll}
 import com.pengrad.telegrambot.model.{MessageEntity, Update}
 import com.pengrad.telegrambot.request.{SendMessage, SendPhoto}
